// src/app/core/models/assetDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface AssetDto {
  id?: string;
  name?: string;
  description?: string;
  assetType?: string;
  purchaseDate?: string;
  purchasePrice?: number;
  currentValue?: number;
  location?: string;
  serialNumber?: string;
  status?: string;
  tenantId?: string;
}