// src/app/core/models/createAssetDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface CreateAssetDto {
  tenantId: string;
  name: string;
  description: string;
  assetType: string;
  identificationNumber: string;
  make: string;
  model: string;
  year: string;
  quantity: string;
  status: string;
  currentLocation: string;
  requiresInspection: boolean;
  inspectionCheckpointsJson: string;
  lastMaintenanceDate: string;
  nextMaintenanceDate: string;
  purchaseDate: string;
  purchaseCost: number;
  conditionNotes: string;
}