// src/app/core/models/assetCheckoutDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface AssetCheckoutDto {
  id: string;
  tenantId: string;
  assetId: string;
  userId: string;
  checkoutDate: string;
  returnDate: string;
  createdAt: string;
  updatedAt: string;
  createdBy: string;
  updatedBy: string;
}