// src/app/core/models/checkoutAssetDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface CheckoutAssetDto {
  assetId: string;
  userId: string;
  checkoutDate: string;
  returnDate: string;
  purpose: string;
  destination: string;
  expectedReturnDate: string;
  startingOdometer: number;
  fuelLevelCheckout: number;
  checkoutNotes: string;
}