// src/app/core/models/checkinAssetDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface CheckinAssetDto {
  assetId: string;
  userId: string;
  returnDate: string;
  conditionNotes: string;
  endingOdometer: number;
  fuelLevelCheckin: number;
  returnedInGoodCondition: boolean;
  damageReported: boolean;
  checkinNotes: string;
  checkoutId: string;
}