// src/app/core/models/assetType.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export type AssetType = typeof AssetType[keyof typeof AssetType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AssetType = {
  Vehicle: 0,
  Tent: 1,
  Equipment: 2,
  Refrigeration: 3,
  Furniture: 4,
  Tools: 5,
  Other: 6,
} as const;